requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
pymongo==4.3.3
python-dotenv==1.0.0
bcrypt==4.0.1
PyPDF2==3.0.1
email-validator==2.0.0
python-multipart==0.0.6
Werkzeug==2.3.7
Pillow==10.0.1
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
